# raptor: bound before launching the reader