This is a corrected version of @antoine-de's PR

benchmark_full, unlimited snd pass, -i 10:
 - without : 197s
 - with : 147s

Still 10 seconds on the very problematic journey.

4s -> 900ms to go to la fourche (journey in jira's ticket)