@antoine-de 2015-11-23
OK for me

Can you add figures about the performance ?