@TeXitoi 2015-11-25
I still have to check the impact of "raptor solution reader: limit the numbers of allocations"